tests/wpt/web-platform-tests/css/css-flexbox-1/getcomputedstyle/flexbox_computedstyle_justify-content-center.html

<!DOCTYPE html>
<title>flexbox | computed style | justify-content: center</title>
<link rel="author" href="http://opera.com" title="Opera Software">
<link rel="help" href="http://www.w3.org/TR/css-flexbox-1/#propdef-justify-content">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="flags" content="dom">
<style>
body {
	color: red;
	justify-content: center;
}
</style>
<div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("justify-content"),
	          "center");
});
</script>
